+++ web/modules/custom/ai_screening/custom_event_dispatcher/src/Event/UserinfoSaveEvent.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Drupal\openid_connect_event_dispatcher\Event;
+namespace Drupal\custom_event_dispatcher\Event;
 
 use Drupal\Component\EventDispatcher\Event;
+use Drupal\custom_event_dispatcher\OpenIdConnectHookEvents;
 use Drupal\hook_event_dispatcher\Attribute\HookEvent;
 use Drupal\hook_event_dispatcher\Event\EventInterface;
-use Drupal\openid_connect_event_dispatcher\OpenIdConnectHookEvents;
 use Drupal\user\UserInterface;
 
 /**
- * UserinfoSaveEvent.
+ * Userinfo save event.
  */
 #[HookEvent(id: 'openid_connect_userinfo_save', hook: 'openid_connect_userinfo_save')]
 class UserinfoSaveEvent extends Event implements EventInterface {